requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
python-multipart>=0.0.9,<1.0
httpx>=0.27,<1.0
PyJWT>=2.9,<3.0
SQLAlchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
alembic>=1.14,<2.0
celery[redis]>=5.4,<6.0
redis>=5.0,<6.0
Authlib>=1.3,<2.0
itsdangerous>=2.2,<3.0
prometheus-client>=0.21,<1.0
opentelemetry-api>=1.29,<2.0
opentelemetry-sdk>=1.29,<2.0
opentelemetry-exporter-otlp-proto-http>=1.29,<2.0
opentelemetry-instrumentation-fastapi>=0.50b0,<1.0
opentelemetry-instrumentation-sqlalchemy>=0.50b0,<1.0
numpy>=1.26,<3.0
pandas>=2.2,<3.0
scikit-learn>=1.4,<2.0
joblib>=1.3,<2.0
Pillow>=10.3,<12.0
pydicom>=3.0,<4.0
vercel>=0.5,<1.0
pytest>=8.2,<9.0
